Remove ghost code
