+++ Assets/ar.lproj/Localizable.strings
@@ -164,7 +164,12 @@
 
 "Cancel all notifications"="إلغاء جميع الإشعارات";
 
+"Notifications set for the entire course"="تم تعيين الإشعارات للدورة بأكملها";
+
+"Notification set for the event"="تم تعيين إشعار للحدث";
+
 "Cancel"="إلغاء";
+
 "No available bookmarks"="لا توجد علامات مرجعية متاحة";
 
 "It looks like there's no available bookmarks"="يبدو أنه لا توجد علامات مرجعية متاحة";
