+++ Assets/bs-BA.lproj/Localizable.strings
@@ -164,6 +164,10 @@
 
 "Cancel all notifications"="Otkažite sva obavještenja";
 
+"Notifications set for the entire course"="Obavijesti postavljene za cijeli kurs";
+
+"Notification set for the event"="Obavijest je postavljena za događaj";
+
 "No available bookmarks"="Nema dostupnih oznaka";
 
 "It looks like there's no available bookmarks"="Čini se da nema dostupnih oznaka";
