+++ Assets/de.lproj/Localizable.strings
@@ -164,6 +164,10 @@
 
 "Cancel all notifications"="Alle benachrichtigungen abstellen";
 
+"Notifications set for the entire course"="Benachrichtigungen für den gesamten Kurs festgelegt";
+
+"Notification set for the event"="Benachrichtigung für das Ereignis wurde gesetzt";
+
 "No available bookmarks"="Keine verfügbaren Lesezeichen";
 
 "It looks like there's no available bookmarks"="Anscheinend sind keine Lesezeichen verfügbar";
