+++ Assets/en.lproj/Localizable.strings
@@ -164,6 +164,10 @@
 
 "Cancel all notifications"="Cancel all notifications";
 
+"Notifications set for the entire course"="Notifications set for the entire course";
+
+"Notification set for the event"="Notification set for the event";
+
 "Cancel"="Cancel";
 
 "No available bookmarks"="No available bookmarks";
