+++ Assets/fr.lproj/Localizable.strings
@@ -164,6 +164,10 @@
 
 "Cancel all notifications"="Annuler toutes les notifications";
 
+"Notifications set for the entire course"="Notifications définies pour l'ensemble du cours";
+
+"Notification set for the event"="Notification définie pour l'événement";
+
 "No available bookmarks"="Aucun favoris disponible";
 
 "It looks like there's no available bookmarks"="Il semble qu'il n'y ait pas de favoris disponibles";
