+++ Assets/it.lproj/Localizable.strings
@@ -164,6 +164,10 @@
 
 "Cancel all notifications"="Annulla tutte le notifiche";
 
+"Notifications set for the entire course"="Notifiche impostate per l'intero corso";
+
+"Notification set for the event"="Notifica impostata per l'evento";
+
 "No available bookmarks"="Nessun programma disponibile";
 
 "It looks like there's no available bookmarks"="Sembra che non ci sono programmi disponibili";
