--- Tumble/Data/Constants/NetworkConstants.swift
@@ -1,12 +1,0 @@
-//
-//  NetworkConstants.swift
-//  Tumble
-//
-//  Created by Adis Veletanlic on 2023-02-17.
-//
-
-import Foundation
-
-class NetworkConstants {
-    static let MAX_CONSECUTIVE_ATTEMPTS: Int = 2
-}
