--- Tumble/Domain/Enums/Status/RootViewStatus.swift
@@ -1,13 +1,0 @@
-//
-//  RootViewStatus.swift
-//  Tumble
-//
-//  Created by Adis Veletanlic on 2023-04-01.
-//
-
-import Foundation
-
-enum RootViewStatus {
-    case onboarding
-    case app
-}
