+++ Tumble/Domain/Model/Network/Nullable.swift
@@ -24,7 +24,9 @@ enum Nullable {
         public required init(from decoder: Decoder) throws {
             let container = try decoder.singleValueContainer()
             if !container.decodeNil() {
-                throw DecodingError.typeMismatch(JSONNull.self, DecodingError.Context(codingPath: decoder.codingPath, debugDescription: "Wrong type for JSONNull"))
+                throw DecodingError.typeMismatch(
+                    JSONNull.self,
+                    DecodingError.Context(codingPath: decoder.codingPath, debugDescription: "Wrong type for JSONNull"))
             }
         }
