+++ Tumble/Domain/Model/Notifications/Notifications+Types.swift
