--- Tumble/Domain/Notifications/NotificationManagerProtocol.swift
@@ -1,29 +1,0 @@
-//
-//  NotificationManagerProtocol.swift
-//  Tumble
-//
-//  Created by Adis Veletanlic on 2023-02-03.
-//
-
-import Foundation
-
-protocol NotificationManagerProtocol {
-    func scheduleNotification(
-        for notification: EventNotification,
-        type: NotificationType,
-        userOffset: Int,
-        completion: @escaping (Result<Int, NotificationError>) -> Void
-    )
-    
-    func cancelNotification(for eventId: String)
-    
-    func cancelNotifications(with categoryIdentifier: String)
-    
-    func cancelNotifications() -> Void
-    
-    func isNotificationScheduled(categoryIdentifier: String, completion: @escaping (Bool) -> Void) -> Void
-    
-    func isNotificationScheduled(eventId: String, completion: @escaping (Bool) -> Void) -> Void
-    
-    func rescheduleEventNotifications(previousOffset: Int, userOffset: Int) -> Void
-}
