+++ Tumble/Extensions/Presentation/View/ImageExtensions.swift
@@ -9,11 +9,6 @@ import Foundation
 import SwiftUI
 
 extension Image {
-    func tabBarIcon(isSelected: Bool) -> some View {
-        font(.system(size: 20, weight: .medium))
-            .foregroundColor(isSelected ? .primary : .onSurface.opacity(0.5))
-            .padding(.bottom, 5)
-    }
     
     func actionIcon() -> some View {
         font(.system(size: 17, weight: .medium))
