+++ Tumble/Observables/Controllers/AppController.swift
@@ -8,10 +8,10 @@
 import Foundation
 import SwiftUI
 
-// Observable state from views to open sheets from
-// pressing external notifications in order to navigate
-// to certain views with attributes. These values are global
-// and meant to be controllable from anywhere in the app.
+/// Observable state from views to open sheets from
+/// pressing external notifications in order to navigate
+/// to certain views with attributes. These values are global
+/// and meant to be controllable from anywhere in the app.
 class AppController: ObservableObject {
     static let shared: AppController = .init()
