+++ Tumble/Observables/ViewModels/EventDetailsSheetViewModel.swift
@@ -108,7 +108,6 @@ final class EventDetailsSheetViewModel: ObservableObject {
                     DispatchQueue.main.async { [weak self] in
                         self?.isNotificationSetForCourse = true
                         self?.isNotificationSetForEvent = true
-                        // TODO: Display success popup
                     }
                 }
             } catch {
@@ -122,6 +121,48 @@ final class EventDetailsSheetViewModel: ObservableObject {
             return try await notificationManager.notificationsAreAllowed()
         } catch {
             return false
+        }
+    }
+    
+    @MainActor
+    func applyNotificationForScheduleEventsInCourse(events: [Event]) async throws -> Bool {
+        for event in events {
+            if let notification = notificationManager.createNotificationFromEvent(event: event) {
+                try await notificationManager.scheduleNotification(
+                    for: notification,
+                    type: .event,
+                    userOffset: notificationOffset
+                )
+                AppLogger.shared.debug("Set notification for \(event.title)")
+            } else {
+                AppLogger.shared.error("Permission error: Not allowed")
+                return false
+            }
+        }
+        return true
+    }
+
+    @MainActor
+    func checkNotificationIsSetForCourse() {
+        if let course = event.course {
+            let courseId = course.courseId
+            Task {
+                let result = await notificationManager.isNotificationScheduled(categoryIdentifier: courseId)
+                DispatchQueue.main.async {
+                    self.isNotificationSetForCourse = result
+                }
+            }
+        }
+    }
+    
+    @MainActor
+    func checkNotificationIsSetForEvent() {
+        let eventId = event.eventId
+        Task {
+            let result = await notificationManager.isNotificationScheduled(eventId: eventId)
+            DispatchQueue.main.async {
+                self.isNotificationSetForEvent = result
+            }
         }
     }
