+++ Tumble/Presentation/Views/Home/Available/TodaysEventsCarousel.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 struct TodaysEventsCarousel: View {
     @Binding var swipedCards: Int
-    @Binding var weekEventCards: [DayEventCardModel]
+    @Binding var weekEventCards: [WeekEventCardModel]
     
     var body: some View {
         ZStack {
@@ -94,7 +94,7 @@ struct TodaysEventsCarousel: View {
 private struct CarouselCard: View {
     let event: Event
     let index: Int
-    let eventsForToday: [DayEventCardModel]
+    let eventsForToday: [WeekEventCardModel]
     @Binding var swipedCards: Int
     
     var body: some View {
